Move ComponentHandlers class to module
