+++ kpops/pipeline_generator/pipeline.py
@@ -13,8 +13,8 @@ from rich.console import Console
 from rich.syntax import Syntax
 
 from kpops.cli.pipeline_config import PipelineConfig
-from kpops.cli.pipeline_handlers import ComponentHandlers
 from kpops.cli.registry import Registry
+from kpops.component_handlers import ComponentHandlers
 from kpops.components.base_components.base_defaults_component import update_nested_pair
 from kpops.components.base_components.pipeline_component import PipelineComponent
 from kpops.utils.yaml_loading import load_yaml_file, substitute
