+++ tests/cli/test_handlers.py
@@ -4,7 +4,7 @@ from pytest_mock import MockerFixture
 
 from kpops.cli.main import setup_handlers
 from kpops.cli.pipeline_config import PipelineConfig
-from kpops.cli.pipeline_handlers import ComponentHandlers
+from kpops.component_handlers import ComponentHandlers
 from kpops.component_handlers.kafka_connect.handler import ConnectorHandler
 from kpops.component_handlers.schema_handler.schema_handler import SchemaHandler
 from kpops.component_handlers.streams_bootstrap.handler import AppHandler
