+++ tests/components/test_pipeline_component.py
@@ -3,7 +3,7 @@ from unittest.mock import MagicMock
 
 from kpops.cli import pipeline_config
 from kpops.cli.pipeline_config import TopicNameConfig
-from kpops.cli.pipeline_handlers import ComponentHandlers
+from kpops.component_handlers import ComponentHandlers
 from kpops.components.base_components.models.to_section import (
     OutputTopicTypes,
     TopicConfig,
