+++ tests/pipeline/test_pipeline.py
@@ -7,7 +7,7 @@ from pytest_mock import MockerFixture
 from typer.testing import CliRunner
 
 from kpops.cli.main import app
-from kpops.cli.pipeline_handlers import ComponentHandlers
+from kpops.component_handlers import ComponentHandlers
 from kpops.utils.yaml_loading import load_yaml_file
 
 runner = CliRunner()
@@ -17,7 +17,7 @@ PIPELINE_BASE_DIR = str(RESOURCE_PATH.parent)
 
 
 @pytest.fixture(autouse=True)
-def pipeline_handlers(mocker: MockerFixture):
+def component_handlers(mocker: MockerFixture):
     def mock_setup_handlers(*args, **kwargs) -> ComponentHandlers:
         return ComponentHandlers(
             schema_handler=None,
